src/pyproject.toml

[tool.poetry]
name = "BD-COUCH-Msc"
version = "0.1.0"
description = "Your project description"
authors = ["Your Name <your@email.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
setuptools = "*"
numpy = "1.24.3"
opencv-python = "*"
matplotlib = "*"
pyglet = "*"
tensorflow="2.11.0"
tensorflow-io-gcs-filesystem = "0.23.1"
tensorboard="*"
tensorboard-data-server="*"
tensorboard-plugin-wit= "*"
tensorflow-estimator= "*"
pandas= "*"
h5py = "*"
tensorlayer = "^2.2.5"
pygame = "^2.5.2"
pyyaml = "^6.0.1"
rospkg = "^1.5.0"
spatialmath-python = "^1.1.8"
pynput = "^1.7.6"
tensorflow-probability = "0.19.0"
tabulate = "^0.9.0"
mujoco = "^2.3.7"
pymunk = "^6.5.2"
shapely = "^2.0.2"
cvxpy = "^1.4.2"
robosuite = "^1.4.1"
cython = "3.0.5"
ipdb = "^0.13.13"
gin-config = "^0.5.0"
torch = "1.12.1+cu113"
diffusers = "^0.32.2"
einops = "0.4.1"
mpi4py = "^4.0.0"
torchvision = "0.13.1"
zarr = ">=2.0, <3.0"
robomimic = "0.3.0"
hydra-core = "^1.3.2"


[tool.poetry.dev-dependencies]
gymnasium = { version = "*"}
gym = { version = "0.23.1"}

[[tool.poetry.source]]
name = "PyPI"
priority = "default"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu113"
priority = "supplemental"

[[tool.poetry.source]]
name = "pytorch3d"
url  = "https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py38_cu113_pyt1110/download.html"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

